Two high level parsing routines are added.
The first one gets console stream messages from the parse tree, and outputs them
as a usual string.

The second one is related to getting frame information. In gdb/mi sense, some
commands bring asynchronous responses because of their nature. For example,
most execution control commands are belong to this category. That information
is taken from exec async record. With this knowledge, it is possible to
implement a simple gdbvim execution environment.