gdbvim ouput is established.
Currently, our approach is to run gdb as it is and to sit between user and gdb
so as to catch and send some relevant commands to the mi interpreter instead of
the console interpreter. For example, program execution commands like next,
continue, step etc. are a good candidate for interception.

There are two sides to be considered: how the input and output will be handled?
For the output side, we have to switch between gdb/cli and gdb/mi accordingly
depending on the given command. In a similar way, we need to find a way to spot
a gdb/mi command from gdb/cli commands. With this commit, we have installed the
output side.