gdbvim input is established.
There are two consumers for input: program and gdb.

A program may need to get input from stdin. This is only possible when the
program is executed. It is handled now.

There might be two different situations for gdb to claim stdin: when an inferior
is stopped and running. The first one is obvious, gdb gets input when program is
stopped. The second one is more subtle. A bit background: gdb console supports
getting commands while it is executing an inferior provided that the inferior
does not want it. Simply, gdb queues commands and when it gets control, executes
them respectively. While the former is supported, the latter has not been
supported yet. In the current situation, all the input while a program is
running are directed to it whether it is interested in or not.